// client/src/app/models/fit-form-user.model.ts
export interface FitFormUser {
    firstName: string;
    lastName: string;
    email: string;
    mobile: number;
    weight: number;
    height: number;
    bmi: number;
    bmiResult: string;
    gender: string;
    requireTrainer: string;
    package: string;
    important: string[];
    haveGymBefore: string;
    enquiryDate: string;
}